Add finished automatic dice roller
This dice roller differs from the one in the tutorial in that if the roll is a draw (if both dice rolls are the same), then it will keep rolling the dice until the two values are different.
I also changed the default font size of the h1 element from the 8rem default to a more-bearable 4rem size.